+++ Changelog.md
@@ -13,6 +13,7 @@
 - Fix: The Codebreaker does not work on keypad doors
 - Fix: The Scanner Door can sometimes be opened/closed by looking at a side the scanner texture is not on
 - Fix: The Portable Tune Player plays the saved tune slightly slower than the Sonic Security System
+- Fix: Reinforced Hoppers pushed by Reinforced Pistons lose their owner
 - Potential Fix: Entities sometimes disappear when dismounting a camera
 
 --------------------------Changelog for v1.9.0.2 of SecurityCraft--------------------------